Refactor creation of grpc servers to allow for controller tests to use it
