+++ pkg/controller/subscription/watcher_test.go
@@ -7,118 +7,16 @@ package subscription
 import (
 	"context"
 	subapi "github.com/onosproject/onos-api/go/onos/e2sub/subscription"
-	subtaskapi "github.com/onosproject/onos-api/go/onos/e2sub/task"
-	e2sub "github.com/onosproject/onos-e2sub/pkg/northbound/subscription"
-	e2task "github.com/onosproject/onos-e2sub/pkg/northbound/task"
 	server2 "github.com/onosproject/onos-e2t/pkg/southbound/e2ap101/server"
 	"github.com/onosproject/onos-lib-go/pkg/controller"
 	"sync"
 
-	substore "github.com/onosproject/onos-e2sub/pkg/store/subscription"
-	taskstore "github.com/onosproject/onos-e2sub/pkg/store/task"
-
-	"github.com/onosproject/onos-lib-go/pkg/northbound"
 	"github.com/stretchr/testify/assert"
-	"google.golang.org/grpc"
-	"google.golang.org/grpc/test/bufconn"
-	"net"
 	"testing"
 )
 
-const E2NodeID = "e2node"
-
-var (
-	lis       = bufconn.Listen(1024 * 1024)
-	server    = grpc.NewServer()
-	taskStore taskstore.Store
-	subStore  substore.Store
-
-	subTask = &subtaskapi.SubscriptionTask{
-		ID:             "1",
-		Revision:       0,
-		SubscriptionID: "1",
-		EndpointID:     E2NodeID,
-		Lifecycle:      subtaskapi.Lifecycle{},
-	}
-)
-
-func bufDialer(context.Context, string) (net.Conn, error) {
-	return lis.Dial()
-}
-
-func newTestSubService(t *testing.T) northbound.Service {
-	var err error
-	subStore, err = substore.NewLocalStore()
-	assert.NoError(t, err)
-
-	return e2sub.NewService(subStore)
-}
-
-func newTestTaskService(t *testing.T) northbound.Service {
-	var err error
-	taskStore, err = taskstore.NewLocalStore()
-	assert.NoError(t, err)
-
-	return e2task.NewService(taskStore)
-}
-
-func createSubServerConnection(t *testing.T) *grpc.ClientConn {
-	lis = bufconn.Listen(1024 * 1024)
-	s := newTestSubService(t)
-	assert.NotNil(t, s)
-
-	s.Register(server)
-
-	ctx := context.Background()
-	conn, err := grpc.DialContext(ctx, "bufnet", grpc.WithContextDialer(bufDialer), grpc.WithInsecure())
-	if err != nil {
-		t.Fatalf("Failed to dial bufnet: %v", err)
-	}
-	return conn
-}
-
-func createTaskServerConnection(t *testing.T) *grpc.ClientConn {
-	s := newTestTaskService(t)
-	assert.NotNil(t, s)
-
-	s.Register(server)
-
-	ctx := context.Background()
-	conn, err := grpc.DialContext(ctx, "bufnet", grpc.WithContextDialer(bufDialer), grpc.WithInsecure())
-	if err != nil {
-		t.Fatalf("Failed to dial bufnet: %v", err)
-	}
-	return conn
-}
-
-func serve(t *testing.T) {
-	go func() {
-		if err := server.Serve(lis); err != nil {
-			assert.NoError(t, err, "Server exited with error: %v", err)
-		}
-	}()
-}
-
-func createServers(t *testing.T) (*grpc.ClientConn, *grpc.ClientConn) {
-	server = grpc.NewServer()
-	subConn := createSubServerConnection(t)
-	taskConn := createTaskServerConnection(t)
-	serve(t)
-	return subConn, taskConn
-}
-
-func createClients(t *testing.T) (subapi.E2SubscriptionServiceClient, subtaskapi.E2SubscriptionTaskServiceClient) {
-	subConn, taskConn := createServers(t)
-
-	subscriptionClient := subapi.NewE2SubscriptionServiceClient(subConn)
-	assert.NotNil(t, subscriptionClient)
-
-	subscriptionTaskClient := subtaskapi.NewE2SubscriptionTaskServiceClient(taskConn)
-	assert.NotNil(t, subscriptionTaskClient)
-	return subscriptionClient, subscriptionTaskClient
-}
-
 func TestWatcher(t *testing.T) {
+	createServerScaffolding(t)
 	_, subscriptionTaskClient := createClients(t)
 
 	watch := Watcher{
@@ -131,7 +29,7 @@ func TestWatcher(t *testing.T) {
 	err := watch.Start(ch)
 	assert.NoError(t, err)
 
-	err = taskStore.Create(context.Background(), subTask)
+	err = scaffold.taskStore.Create(context.Background(), subTask)
 	assert.NoError(t, err)
 
 	newID := <-ch
@@ -141,6 +39,7 @@ func TestWatcher(t *testing.T) {
 }
 
 func TestChannelWatcher(t *testing.T) {
+	createServerScaffolding(t)
 	subscriptionClient, subscriptionTaskClient := createClients(t)
 
 	watch := ChannelWatcher{
@@ -156,7 +55,7 @@ func TestChannelWatcher(t *testing.T) {
 	err := watch.Start(ch)
 	assert.NoError(t, err)
 
-	err = taskStore.Create(context.Background(), subTask)
+	err = scaffold.taskStore.Create(context.Background(), subTask)
 	assert.NoError(t, err)
 
 	subscription := &subapi.Subscription{
